#1028 Regression: Aspect on duplicate declaration causes failure in the linker.
