+++ Metalama.Framework/src/tests/Metalama.Framework.Tests.AspectTests/Tests/Aspects/InvalidCode/DuplicateMethod.cs
@@ -2,22 +2,21 @@
 // SharpCrafters s.r.o. licenses this file to you under either the MIT license or a proprietary license, depending on the repository from which it was obtained.
 // Refer to LICENSE.md in the repository root for complete details.
 
-#if TESTOPTIONS
-// @Skipped(#1028)
-#endif
-
+using System;
 using Metalama.Framework.Aspects;
 
-namespace Metalama.Framework.Tests.PublicPipeline.Aspects.InvalidCode.DuplicateDeclarationWithAspect;
+namespace Metalama.Framework.Tests.PublicPipeline.Aspects.InvalidCode.DuplicateMethod;
 
 /*
- * Tests that when there are duplicate declarations, the error is produced without crashing.
+ * Tests that ambiguous declaration does not cause a crash in the linker. The output may not be correct.
  */
 
-public class Aspect : OverrideMethodAspect
+internal class Aspect : OverrideMethodAspect
 {
     public override dynamic? OverrideMethod()
     {
+        Console.WriteLine( "Aspect" );
+
         return meta.Proceed();
     }
 }
@@ -26,16 +25,15 @@ public class Aspect : OverrideMethodAspect
 internal class TargetCode
 {
     [Aspect]
-    public int Foo()
+    private int Method( int a )
     {
-        return 42;
+        return a;
     }
 
 #if TESTRUNNER
-    [Aspect]
-    public int Foo()
+    int Method(int a)
     {
-        return 42;
+        return a;
     }
 #endif
 }
