+++ Metalama.Framework/src/tests/Metalama.Framework.Tests.AspectTests/Tests/Aspects/InvalidCode/DuplicateMethodWithAspect.cs
@@ -2,22 +2,18 @@
 // SharpCrafters s.r.o. licenses this file to you under either the MIT license or a proprietary license, depending on the repository from which it was obtained.
 // Refer to LICENSE.md in the repository root for complete details.
 
-using System;
-using Metalama.Framework.Advising;
 using Metalama.Framework.Aspects;
 
-namespace Metalama.Framework.Tests.PublicPipeline.Aspects.InvalidCode.DuplicateDeclaration;
+namespace Metalama.Framework.Tests.PublicPipeline.Aspects.InvalidCode.DuplicateMethodWithAspect;
 
 /*
- * Tests that ambiguous declaration does not cause a crash in the linker. The output may not be correct.
+ * Tests that when there are duplicate declarations, the error is produced without crashing.
  */
 
-internal class Aspect : OverrideMethodAspect
+public class Aspect : OverrideMethodAspect
 {
     public override dynamic? OverrideMethod()
     {
-        Console.WriteLine( "Aspect" );
-
         return meta.Proceed();
     }
 }
@@ -26,15 +22,16 @@ internal class Aspect : OverrideMethodAspect
 internal class TargetCode
 {
     [Aspect]
-    private int Method( int a )
+    public int Foo()
     {
-        return a;
+        return 42;
     }
 
 #if TESTRUNNER
-    int Method(int a)
+    [Aspect]
+    public int Foo()
     {
-        return a;
+        return 42;
     }
 #endif
 }
