+++ Metalama.Framework/src/tests/Metalama.Framework.Tests.AspectTests/Tests/Aspects/InvalidCode/DuplicatePropertyWithAspect.cs
@@ -1,0 +1,48 @@
+// Copyright (c) 2020-2025 SharpCrafters s.r.o. and contributors.
+// SharpCrafters s.r.o. licenses this file to you under either the MIT license or a proprietary license, depending on the repository from which it was obtained.
+// Refer to LICENSE.md in the repository root for complete details.
+
+#if TESTOPTIONS
+// @Skipped(#1056)
+#endif
+
+using Metalama.Framework.Aspects;
+using System;
+
+namespace Metalama.Framework.Tests.PublicPipeline.Aspects.InvalidCode.DuplicatePropertyWithAspect;
+
+/*
+ * Tests that when there are duplicate declarations, the error is produced without crashing.
+ */
+
+internal class Aspect : OverrideFieldOrPropertyAspect
+{
+    public override dynamic? OverrideProperty
+    {
+        get
+        {
+            Console.WriteLine( "Aspect" );
+
+            return meta.Proceed();
+        }
+
+        set
+        {
+            Console.WriteLine( "Aspect" );
+
+            meta.Proceed();
+        }
+    }
+}
+
+// <target>
+internal class TargetCode
+{
+    [Aspect]
+    private int Property { get; set; }
+
+#if TESTRUNNER
+    [Aspect]
+    private int Property { get; set; }
+#endif
+}
